rename the folders
