+++ game/directing/director.py
@@ -3,7 +3,7 @@ from game.casting.rock import Rock
 
 from game.shared.color import Color
 from game.shared.point import Point
-from game.shared.constants import Constants
+from game.shared.constant import Constants
 import random
 
 constants = Constants()
