+++ game/shared/point.py
